+++ generator/src/base_type_generator.ts
@@ -3,6 +3,7 @@ import { ImportAlias } from './constants'
 import { GenerateContext } from './generate_context'
 import { GeneratorBase } from './generator_base'
 import { NexemaFile, NexemaTypeDefinition } from './models'
+import { getJavascriptType } from './utils'
 
 export class BaseTypeGenerator extends GeneratorBase {
     public constructor(type: NexemaTypeDefinition, file: NexemaFile, context: GenerateContext) {
@@ -22,7 +23,7 @@ export class BaseTypeGenerator extends GeneratorBase {
         let output = ''
 
         for (const field of this._type.fields!) {
-            const jsType = this.getJavascriptType(field.type!)
+            const jsType = getJavascriptType(this._context, this._file, field.type!)
             output += `${this._writeHeader(this._type.documentation, this._type.annotations, true)}
             public abstract get ${this._fieldNames[field.name]}(): ${jsType};
             ${this._writeHeader(this._type.documentation, this._type.annotations, true)}
